refactor(card-inspector): split handlers.py into content/printing-nav/image-pipeline mixins (#805)
Behavior-preserving decomposition of the 463-LOC
widgets/panels/card_inspector_panel/handlers.py into a handlers/ topic
sub-package, mirroring widgets/frames/identify_opponent/handlers/.

- handlers/__init__.py: composes CardInspectorPanelHandlersMixin from the
  three topic mixins and re-exports it, so frame.py's import is unchanged.
- handlers/content.py: ContentMixin — reset()/update_card() population plus
  the set_* dependency setters and _render_mana_cost().
- handlers/printing_nav.py: PrintingNavMixin — prev/next/autosave/save
  callbacks, _current_printing_dict, _emit_printing_changed/_selected,
  _set_printing_label.
- handlers/image_pipeline.py: ImagePipelineMixin — async lookup, UI-thread
  appliers, inbound download/printings callbacks, _set_display_mode,
  _notify_selection, _request_missing_image.

protocol.py gains the cross-mixin attrs the split now exposes to the Protocol
base (inspector_selection, _printing_selected_handler, _autosave_printing,
save_panel, autosave_checkbox, save_art_btn).

Methods are moved verbatim; no runtime behavior, public API, or wx event
wiring changed. _resolve_image_request_name/_request_matches_current/
_failure_key remain in properties.py.
